api/requirements.txt
uvicorn
fastapi
typing
pillow
# torch
# #torch==1.12.0+cu113
# torchvision
#torchvision==0.13.0+cu113
python-multipart
opencv-python-headless
face-detection==0.2.2
pandas
PyYAML==6.0
tqdm
matplotlib
seaborn

#yolov5 dependencies https://github.com/ultralytics/yolov5
# Base ----------------------------------------
numpy>=1.18.5
requests>=2.23.0
protobuf<4.21.3  # https://github.com/ultralytics/yolov5/issues/8012

# Logging -------------------------------------
tensorboard>=2.4.1
# wandb

# Plotting ------------------------------------
# pandas>=1.1.4
# seaborn>=0.11.0

# Export --------------------------------------
# coremltools>=4.1  # CoreML export
# onnx>=1.9.0  # ONNX export
# onnx-simplifier>=0.3.6  # ONNX simplifier
# scikit-learn==0.19.2  # CoreML quantization
# tensorflow>=2.4.1  # TFLite export
# tensorflowjs>=3.9.0  # TF.js export
# openvino-dev  # OpenVINO export

# Extras --------------------------------------
ipython  # interactive notebook
psutil  # system utilization
thop  # FLOPs computation
# albumentations>=1.0.3
# pycocotools>=2.0  # COCO mAP
# roboflow